Cambie long por integer en el repositorio de soporte, ya que estaba definido como integer, no como long, y estoy terminando el SoporteServiceTest, me falta un poquito

+++ src/main/java/com/ecomarket/cl/ecomarket/repository/SoporteRepository.java
@@ -3,5 +3,5 @@ package com.ecomarket.cl.ecomarket.repository;
 import com.ecomarket.cl.ecomarket.model.Soporte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SoporteRepository extends JpaRepository<Soporte, Long> {
+public interface SoporteRepository extends JpaRepository<Soporte, Integer> {
 }
